* Fix config paths
