+++ tools/e2e-core/config/playwright.config.default.js
